various edits
Substantial: began working on the instance object.

+++ src/OneWayTrip/Player.java
@@ -14,9 +14,9 @@ public class Player extends Entity {
     private int experience;
     private String classTag;
     //Archetype weights
-    private double brawn,
-                   finesse,
-                   sorcery;
+    private int brawn,
+                finesse,
+                sorcery;
     //Stat Values
     private int tenacity,
                 dexterity,
@@ -42,7 +42,7 @@ public class Player extends Entity {
         dexterity = d;
         arcane  = a;
     }
-    public void setAllArch(double b, double f, double s){
+    public void setAllArch(int b, int f, int s){
         brawn = b;
         finesse = f;
         sorcery = s;
@@ -72,13 +72,13 @@ public class Player extends Entity {
     public String getTag(){
         return classTag;
     }
-    public double getBra(){
+    public int getBra(){
         return brawn;
     }
-    public double getFin(){
+    public int getFin(){
         return finesse;
     }
-    public double getSor(){
+    public int getSor(){
         return sorcery;
     }
     public int getTen(){
